requirements.txt
boto3==1.26.0
requests==2.28.1
tweepy==4.12.1
python-dotenv==0.21.0
